+++ R/spatial_stars.R
@@ -51,7 +51,6 @@ tmapReproject.dimensions = function(shp, tmapID, bbox = NULL, ..., crs) {
 
 #' @export
 tmapShape.stars = function(shp, is.main, crs, bbox, unit, filter, layer, shp_name, smeta, o, tmf) {
-	print("shape")
 	if (identical(crs, "auto")) crs = auto_crs(shp, crs_extra = o$crs_extra, crs_global = o$crs_global)
 
 	dev = getOption("tmap.devel.mode")